refactor: test/numeric
