+++ test/test-600-others/sources/tools/numeric/test-staff.hpp
